Refactoring Sudo file modifications
Previously templating was used to update the sudoers file, this meant a changed event was triggered each time the role was run when in reality, after the first run-through, nothing changed.

To correct this the 'lineinfile' module is used to ensure required changes are applied, but not applied again if not needed. Since the modifications to these files was so small (a single line per file) this is fine (and preferable).

This means this role when run a second time will return 'ok' or 'skipped' for all of its tasks. Although these changes seem significant, they don't break backwards compatibility and don't add or change any existing feature (just how they are implemented), therefore these changes are made as a patch release.