Removing Sudoers support
This removes all tasks related to modifying the sudoers file during bootstrapping.

This role now requires the underlying base-box (since this is a vagrant specific bootstrapping role) to implement the required features this role used to ensure directly. I.e. the base box should include these features when it is built, rather than when it is provisioned.

This reduces the complexity of this role significantly (the lineinfile module tasks didn't work correctly anyway). Assuming users are using the BAS maintained base box they will notice no difference (other than needing to rebuild their VMs).

As this is definitely a breaking change, this forms a new major release (2.0.0).